Edição de dados de candidato e empresa, além de melhorias na exibição de vagas e candidaturas.

+++ src/back/todosimple/src/main/java/com/worklink/todosimple/aplicacoes/service/AplicacaoService.java
@@ -2,6 +2,7 @@ package com.worklink.todosimple.aplicacoes.service;
 
 import com.worklink.todosimple.aplicacoes.model.Aplicacao;
 import com.worklink.todosimple.aplicacoes.repositories.AplicacaoRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,7 +20,7 @@ public class AplicacaoService {
     }
 
     // Atualizar o status de uma aplicação existente
-    public Aplicacao atualizarStatus(Integer id, String novoStatus) {
+    public Aplicacao atualizarStatus(Long id, String novoStatus) {
         Aplicacao aplicacao = aplicacaoRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Aplicação não encontrada com o ID: " + id));
         aplicacao.setStatus(novoStatus);
